refactor: 패키지 구조를 역할에 맞게 변경
