+++ src/main/java/chess/domain/board/ChessBoard.java
@@ -6,12 +6,12 @@ import chess.domain.piece.King;
 import chess.domain.piece.Knight;
 import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
-import chess.domain.piece.Pieces;
 import chess.domain.piece.Queen;
 import chess.domain.piece.Rook;
-import chess.domain.piece.TeamColor;
+import chess.domain.result.Pieces;
 import chess.domain.result.Result;
 import chess.domain.result.Score;
+import chess.domain.state.TeamColor;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
