+++ src/main/java/chess/domain/piece/PieceOnBoard.java
@@ -3,6 +3,8 @@ package chess.domain.piece;
 import chess.domain.board.ChessBoard;
 import chess.domain.board.Position;
 import chess.domain.result.Score;
+import chess.domain.state.State;
+import chess.domain.state.TeamColor;
 import java.util.HashSet;
 import java.util.Set;
