+++ src/main/java/chess/domain/state/State.java
@@ -1,4 +1,4 @@
-package chess.domain.piece;
+package chess.domain.state;
 
 public enum State {
