+++ src/main/java/chess/domain/state/TeamColor.java
@@ -1,4 +1,4 @@
-package chess.domain.piece;
+package chess.domain.state;
 
 public enum TeamColor {
     BLACK, WHITE, NONE
