+++ src/test/java/domain/PiecesTest.java
@@ -8,11 +8,11 @@ import chess.domain.piece.King;
 import chess.domain.piece.Knight;
 import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
-import chess.domain.piece.Pieces;
 import chess.domain.piece.Queen;
 import chess.domain.piece.Rook;
-import chess.domain.piece.TeamColor;
+import chess.domain.result.Pieces;
 import chess.domain.result.Score;
+import chess.domain.state.TeamColor;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
