+++ src/test/java/domain/piece/RookTest.java
@@ -10,8 +10,8 @@ import chess.domain.board.Position;
 import chess.domain.piece.Blank;
 import chess.domain.piece.Piece;
 import chess.domain.piece.Rook;
-import chess.domain.piece.State;
-import chess.domain.piece.TeamColor;
+import chess.domain.state.State;
+import chess.domain.state.TeamColor;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
